requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
flask-mail>=0.10.0
sqlalchemy>=2.0.40
werkzeug>=3.1.3
flask-wtf>=1.2.2
python-dotenv>=1.0.0